exercise generate-report in tests

+++ tests/test-examples.sh
@@ -34,6 +34,27 @@ for machine in local titan cori theta; do
         -o test_output/$machine-param_test
 done
 
+# run local param test and analyze results, to exercise report generator
+pushd test_output/local-param_test/$USER
+./run-all.sh
+./test/wait.sh >/dev/null
+cd ..
+$CHEETAH_DIR/cheetah.py generate-report >report.log 2>&1
+rval=$?
+if [ ! -s campaign_results.csv -o $rval != 0 ]; then
+    echo "ERROR: generate-report for local-param_test failed"
+    echo "Return value: $rval"
+    exit 1
+fi
+grep -q 'WARN\|ERR' report.log
+if [ $? = 0 ]; then
+    report_path="$(pwd)"/report.log
+    echo "ERROR: generate-report log has warnings or errors,"\
+         "see '$report_path'"
+    exit 1
+fi
+popd
+
 echo loal-heat-simple
 rm -rf test_output/local-heat-simple/*
 ./cheetah.py create-campaign -e examples/heat_transfer_simple.py -m local \
